requirements.txt
requests>=2.25.0
pandas>=2.0.0
schedule>=1.2.2
python-dotenv>=0.21.0
discord.py>=2.6.0
